backend/requirements.txt
Django==6.0.3
django-cors-headers==4.7.0
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
gunicorn==23.0.0
whitenoise==6.9.0
asgiref==3.11.1
sqlparse==0.5.5
tzdata==2026.1
pyjwt==2.10.1
cryptography==44.0.2
PyPDF2==3.0.1
